Split tests up to avoid large files
